requirements.txt
asgiref==3.7.2
Django==4.0
djangorestframework==3.14.0
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv @ git+https://github.com/theskumar/python-dotenv.git@d6c0b9638349a7dd605d60ee555ff60421c1a594
pytz==2024.1
setuptools==69.1.1
sqlparse==0.4.4
tzdata==2024.1
